+++ packages/access-api/src/bindings.d.ts
@@ -1,11 +1,16 @@
 import type { Logging } from '@web3-storage/worker-utils/logging'
-import type { SpaceTable } from '@web3-storage/access/types'
+import type {
+  AccountTable,
+  DelegationTable,
+  SpaceTable,
+} from '@web3-storage/access/types'
 import type { Handler as _Handler } from '@web3-storage/worker-utils/router'
 import { Email } from './utils/email.js'
 import { Spaces } from './models/spaces.js'
 import { Validations } from './models/validations.js'
 import { loadConfig } from './config.js'
-import { Signer as EdSigner } from '@ucanto/principal/ed25519'
+import { ConnectionView, Signer as EdSigner } from '@ucanto/principal/ed25519'
+import { Accounts } from './models/accounts.js'
 
 export {}
 
@@ -27,10 +32,9 @@ export interface Env {
    * publicly advertised decentralized identifier of the running api service
    * * this may be used to filter incoming ucanto invocations
    */
-  DID: string
+  DID: `did:web:${string}`
   // URLs to upload-api so we proxy invocations to it
   UPLOAD_API_URL: string
-  UPLOAD_API_URL_STAGING: string
   // secrets
   PRIVATE_KEY: string
   SENTRY_DSN: string
@@ -54,11 +58,9 @@ export interface RouteContext {
   models: {
     spaces: Spaces
     validations: Validations
+    accounts: Accounts
   }
-  uploadApi: {
-    production?: URL
-    staging?: URL
-  }
+  uploadApi: ConnectionView
 }
 
 export type Handler = _Handler<RouteContext>
@@ -99,4 +101,6 @@ export interface D1ErrorRaw extends Error {
 
 export interface D1Schema {
   spaces: SpaceTable
+  accounts: AccountTable
+  delegations: DelegationTable
 }