+++ packages/access-api/src/service/upload-api-proxy.js
@@ -16,7 +16,6 @@ import { createProxyHandler } from '../ucanto/proxy.js'
  * @template {string|number|symbol} M
  * @template {Ucanto.ConnectionView<any>} [Connection=Ucanto.ConnectionView<any>]
  * @param {object} options
- * @param {Ucanto.Signer} [options.signer]
  * @param {Array<M>} options.methods
  * @param {{ default: Connection } & Record<Ucanto.UCAN.DID, Connection>} options.connections
  */
@@ -38,10 +37,13 @@ function createProxyService(options) {
  */
 
 /**
+ * Create a ucanto connection to an upload api url.
+ * Assumes upload-api at that URL decodes requests as CAR and encodes responses as CBOR.
+ *
  * @param {UcantoHttpConnectionOptions} options
  * @returns {Ucanto.ConnectionView<any>}
  */
-function createUcantoHttpConnection(options) {
+export function createUploadApiConnection(options) {
   return Client.connect({
     id: DID.parse(options.audience),
     encoder: CAR,
@@ -53,81 +55,30 @@ function createUcantoHttpConnection(options) {
   })
 }
 
-const uploadApiEnvironments = {
-  production: {
-    audience: /** @type {const} */ ('did:web:web3.storage'),
-    url: new URL('https://up.web3.storage'),
-  },
-  staging: {
-    audience: /** @type {const} */ ('did:web:staging.web3.storage'),
-    url: new URL('https://staging.up.web3.storage'),
-  },
-}
-
-/**
- * @typedef {keyof typeof uploadApiEnvironments} UploadApiEnvironmentName
- * @typedef {typeof uploadApiEnvironments[UploadApiEnvironmentName]['audience']} UploadApiAudience
- */
-
 /**
  * @param {object} options
- * @param {typeof globalThis.fetch} [options.fetch]
- * @param {object} options.uploadApi
- * @param {URL} [options.uploadApi.production]
- * @param {URL} [options.uploadApi.staging]
- */
-function getDefaultConnections(options) {
-  const { fetch = globalThis.fetch, uploadApi } = options
-  return {
-    default: createUcantoHttpConnection({
-      ...uploadApiEnvironments.production,
-      ...(uploadApi.production && { url: uploadApi.production }),
-      fetch,
-    }),
-    ...(uploadApi.staging && {
-      [uploadApiEnvironments.staging.audience]: createUcantoHttpConnection({
-        ...uploadApiEnvironments.staging,
-        url: uploadApi.staging,
-        fetch,
-      }),
-    }),
-  }
-}
-
-/**
- * @template {Ucanto.ConnectionView<any>} [Connection=Ucanto.ConnectionView<any>]
- * @param {object} options
- * @param {Ucanto.Signer} [options.signer]
- * @param {typeof globalThis.fetch} [options.fetch]
- * @param {{ default: Connection, [K: Ucanto.UCAN.DID]: Connection }} [options.connections]
- * @param {Record<Ucanto.UCAN.DID, URL>} [options.audienceToUrl]
- * @param {object} options.uploadApi
- * @param {URL} [options.uploadApi.production]
- * @param {URL} [options.uploadApi.staging]
+ * @param {import('../bindings.js').RouteContext['uploadApi']} options.uploadApi
  */
 export function createUploadProxy(options) {
   return createProxyService({
     ...options,
-    connections: options.connections || getDefaultConnections(options),
+    connections: {
+      default: options.uploadApi,
+    },
     methods: ['list', 'add', 'remove', 'upload'],
   })
 }
 
 /**
- * @template {Ucanto.ConnectionView<any>} [Connection=Ucanto.ConnectionView<any>]
  * @param {object} options
- * @param {Ucanto.Signer} [options.signer]
- * @param {typeof globalThis.fetch} [options.fetch]
- * @param {{ default: Connection, [K: Ucanto.UCAN.DID]: Connection }} [options.connections]
- * @param {Record<Ucanto.UCAN.DID, URL>} [options.audienceToUrl]
- * @param {object} options.uploadApi
- * @param {URL} [options.uploadApi.production]
- * @param {URL} [options.uploadApi.staging]
+ * @param {import('../bindings.js').RouteContext['uploadApi']} options.uploadApi
  */
 export function createStoreProxy(options) {
   return createProxyService({
     ...options,
-    connections: options.connections || getDefaultConnections(options),
+    connections: {
+      default: options.uploadApi,
+    },
     methods: ['list', 'add', 'remove', 'store'],
   })
 }