+++ packages/access-api/src/utils/context.js
@@ -1,10 +1,14 @@
 import { Logging } from '@web3-storage/worker-utils/logging'
 import Toucan from 'toucan-js'
+import { Signer } from '@ucanto/principal/ed25519'
 import pkg from '../../package.json'
 import { loadConfig } from '../config.js'
+import { Accounts } from '../models/accounts.js'
 import { Spaces } from '../models/spaces.js'
 import { Validations } from '../models/validations.js'
 import { Email } from './email.js'
+import { createUploadApiConnection } from '../service/upload-api-proxy.js'
+import { DID } from '@ucanto/core'
 
 /**
  * Obtains a route context object.
@@ -16,6 +20,8 @@ import { Email } from './email.js'
  */
 export function getContext(request, env, ctx) {
   const config = loadConfig(env)
+
+  // Sentry
   const sentry = new Toucan({
     context: ctx,
     request,
@@ -30,6 +36,8 @@ export function getContext(request, env, ctx) {
     release: config.VERSION,
     pkg,
   })
+
+  // Logging
   const log = new Logging(request, ctx, {
     token: config.LOGTAIL_TOKEN,
     debug: config.DEBUG,
@@ -42,24 +50,22 @@ export function getContext(request, env, ctx) {
   const url = new URL(request.url)
   return {
     log,
-    signer: config.signer,
+    signer: Signer.parse(config.PRIVATE_KEY).withDID(config.DID),
     config,
     url,
     models: {
       spaces: new Spaces(config.DB),
       validations: new Validations(config.VALIDATIONS),
+      accounts: new Accounts(config.DB),
     },
     email: new Email({
       token: config.POSTMARK_TOKEN,
       sender: config.POSTMARK_SENDER,
     }),
-    uploadApi: {
-      production: config.UPLOAD_API_URL
-        ? new URL(config.UPLOAD_API_URL)
-        : undefined,
-      staging: config.UPLOAD_API_URL_STAGING
-        ? new URL(config.UPLOAD_API_URL_STAGING)
-        : undefined,
-    },
+    uploadApi: createUploadApiConnection({
+      audience: DID.parse(config.DID).did(),
+      url: new URL(config.UPLOAD_API_URL),
+      fetch: globalThis.fetch.bind(globalThis),
+    }),
   }
 }