+++ packages/access-client/src/types.ts
@@ -31,13 +31,16 @@ import type {
   VoucherClaim,
   VoucherRedeem,
   Top,
+  AccessAuthorize,
 } from '@web3-storage/capabilities/types'
 import type { SetRequired } from 'type-fest'
 import { Driver } from './drivers/types.js'
-import type { ColumnType, Selectable } from 'kysely'
+import { SpaceUnknown } from './errors.js'
+import type { ColumnType, Generated, Selectable } from 'kysely'
 
 // export other types
 export * from '@web3-storage/capabilities/types'
+export * from './errors.js'
 
 /**
  * D1 Types
@@ -48,12 +51,32 @@ export interface SpaceTable {
   agent: URI<'did:'>
   email: string
   product: URI<`${string}:`>
-  inserted_at: ColumnType<Date, never, Date>
+  inserted_at: Generated<Date>
   updated_at: ColumnType<Date, never, Date>
   metadata: SpaceTableMetadata | null
   invocation: string
   delegation: string | null
 }
+export type SpaceRecord = Selectable<SpaceTable>
+
+export interface AccountTable {
+  did: URI<'did:'>
+  inserted_at: Generated<Date>
+  updated_at: ColumnType<Date, never, Date>
+}
+export type AccountRecord = Selectable<AccountTable>
+
+export interface DelegationTable {
+  cid: string
+  bytes: Uint8Array
+  audience: URI<'did:'>
+  issuer: URI<'did:'>
+  expires_at: Date | null
+  inserted_at: Generated<Date>
+  updated_at: ColumnType<Date, never, Date>
+}
+
+export type DelegationRecord = Selectable<DelegationTable>
 
 export interface SpaceTableMetadata {
   space: SpaceMeta
@@ -64,6 +87,10 @@ export interface SpaceTableMetadata {
  * Access api service definition type
  */
 export interface Service {
+  access: {
+    // returns a URL string for tests or nothing in other envs
+    authorize: ServiceMethod<AccessAuthorize, string | undefined, Failure>
+  }
   voucher: {
     claim: ServiceMethod<
       VoucherClaim,
@@ -73,7 +100,7 @@ export interface Service {
     redeem: ServiceMethod<VoucherRedeem, void, Failure>
   }
   space: {
-    info: ServiceMethod<SpaceInfo, Selectable<SpaceTable>, Failure>
+    info: ServiceMethod<SpaceInfo, SpaceRecord, Failure | SpaceUnknown>
     'recover-validation': ServiceMethod<
       SpaceRecoverValidation,
       EncodedDelegation<[SpaceRecover]> | undefined,
@@ -179,7 +206,10 @@ export type InvokeOptions<
   A extends Ability,
   R extends Resource,
   CAP extends CapabilityParser<
-    Match<{ can: A; with: R; nb?: {} | undefined }, UnknownMatch>
+    Match<
+      { can: A; with: R; nb?: Record<string, unknown> | undefined },
+      UnknownMatch
+    >
   >
 > = UCANBasicOptions &
   InferNb<InferInvokedCapability<CAP>['nb']> & {
@@ -253,20 +283,21 @@ export interface UCANBasicOptions {
 /**
  * Given an inferred capability infers if the nb field is optional or not
  */
-export type InferNb<C extends {} | undefined> = keyof C extends never
-  ? {
-      nb?: never
-    }
-  : {
-      /**
-       * Non-normative fields for the capability
-       *
-       * Check the capability definition for more details on the `nb` field.
-       *
-       * @see {@link https://github.com/ucan-wg/spec#241-nb-non-normative-fields Spec}
-       */
-      nb: C
-    }
+export type InferNb<C extends Record<string, unknown> | undefined> =
+  keyof C extends never
+    ? {
+        nb?: never
+      }
+    : {
+        /**
+         * Non-normative fields for the capability
+         *
+         * Check the capability definition for more details on the `nb` field.
+         *
+         * @see {@link https://github.com/ucan-wg/spec#241-nb-non-normative-fields Spec}
+         */
+        nb: C
+      }
 
 export interface ClientCodec extends RequestEncoder, ResponseDecoder {}